Add files via upload
This project presents a machine learning project that applies Principal Component Analysis (PCA) and K-Means clustering to e-commerce data for market segmentation analysis. The objective is to discover meaningful patterns in product pricing and customer ratings, thereby unveiling distinct product clusters. The project demonstrates data preprocessing, dimensionality reduction with PCA, and segmentation using K-Means, culminating in a visual representation of the market segments. It serves as a case study in utilizing unsupervised learning for strategic business insights in retail.